Changing to doubles to floats to for native m1 mac support

+++ src/include/_c_mandelbrot_benchmark.h
@@ -2,11 +2,11 @@
 #define _C_MANDELBROT_BENCHMARK_H
 
 // calculate a mandelbrot set
-int _c_mandelbrot(double r, double c, int max_iter, double divergence) {
-  double zr = 0.0;
-  double zi = 0.0;
-  double zr2 = 0.0;
-  double zi2 = 0.0;
+int _c_mandelbrot(float r, float c, int max_iter, float divergence) {
+  float zr = 0.0;
+  float zi = 0.0;
+  float zr2 = 0.0;
+  float zi2 = 0.0;
   int i = 0;
   while (i < max_iter && zr2 + zi2 < divergence) {
     zi = 2.0 * zr * zi + c / 500 - 1;
